implement http connect proxy support
This is lightly based off a patch I found on the rtmpdump list[0], but
then updated to use an HTTP CONNECT proxy approach instead of the
proxy for each POST request. It has only been lightly tested in one
environment and is probably broken in every other situation, as well
as the one I'm actually using it in.

0. http://lists.mplayerhq.hu/pipermail/rtmpdump/2014-April/002355.html